requirements.txt
Flask
psycopg2
requests
python-dotenv